Moved geoserver specific code to geonode.geoserver
