+++ geonode/geoserver/ows.py
@@ -28,6 +28,7 @@ from owslib.coverage.wcsBase import ServiceException
 from owslib.util import http_post
 import urllib
 from geonode import GeoNodeException
+from geonode.geoserver.helpers import ogc_server_settings
 from re import sub
 
 logger = logging.getLogger(__name__)
@@ -132,60 +133,4 @@ def wms_links(wms_url, identifier, bbox, srid, height, width):
         output.append((ext, name, mime, url))
     return output
 
-def wps_execute_layer_attribute_statistics(layer_name, field):
-    """Derive aggregate statistics from WPS endpoint"""
 
-    from geonode.utils import ogc_server_settings
-    # generate statistics using WPS
-    url = '%s/ows' % (ogc_server_settings.LOCATION)
-
-    # TODO: use owslib.wps.WebProcessingService for WPS interaction
-    # this requires GeoServer's WPS gs:Aggregate function to
-    # return a proper wps:ExecuteResponse
-
-
-    request = render_to_string('layers/wps_execute_gs_aggregate.xml', {
-                               'layer_name': 'geonode:%s' %  layer_name,
-                               'field': field
-                              })
-
-    response = http_post(url, request, timeout=ogc_server_settings.TIMEOUT)
-
-    exml = etree.fromstring(response)
-
-    result = {}
-
-    for f in ['Min', 'Max', 'Average', 'Median', 'StandardDeviation', 'Sum']:
-        fr = exml.find(f)
-        if fr is not None:
-            result[f] = fr.text
-        else:
-            result[f] = 'NA'
-
-    count = exml.find('Count')
-    if count is not None:
-        result['Count'] = int(count.text)
-    else:
-        result['Count'] = 0
-
-    result['unique_values'] = 'NA'
-
-    # TODO: find way of figuring out threshold better
-    if result['Count'] < 10000:
-        request = render_to_string('layers/wps_execute_gs_unique.xml', {
-                                   'layer_name': 'geonode:%s' %  layer_name,
-                                   'field': field
-                                  })
-
-        response = http_post(url, request, timeout=ogc_server_settings.TIMEOUT)
-
-        exml = etree.fromstring(response)
-
-        values = []
-
-        for value in exml.findall('{http://www.opengis.net/gml}featureMember/{http://www.opengis.net/gml}UniqueValue/{http://www.opengis.net/gml}value'):
-            if value is not None:
-                values.append(value.text)
-        result['unique_values'] = ','.join(values)
-
-    return result
