+++ geonode/layers/templates/layers/layer_base.html
@@ -12,7 +12,7 @@
       <ul class="nav nav-tabs" id="subnav-tabs">
             <li><a href="{% url "layer_browse" %}" class="explore">{% trans "Explore Layers" %}</a></li>
             <!-- switch to us the new upload -->
-            <li><a href="{% url "data_upload" %}" class="upload">{% trans "Upload Layers" %}</a></li>
+            <li><a href="{% url "layer_upload" %}" class="upload">{% trans "Upload Layers" %}</a></li>
       </ul>
     </div>
     {% block body %}{% endblock body %}
