+++ geonode/layers/templates/upload/layer_upload.html
