+++ geonode/layers/templates/upload/layer_upload_invalid.html
@@ -5,6 +5,6 @@
 {% block body %}
 <div class="twocol">
     <p>{% trans "Your upload is either complete or you haven't resumed an earlier one." %}</p>
-    <p>{% trans "Return to" %} <a href="{% url "data_upload" %}">{% trans "Upload Form" %}</a>.</p>
+    <p>{% trans "Return to" %} <a href="{% url "layer_upload" %}">{% trans "Upload Form" %}</a>.</p>
 </div>
 {% endblock %}
