+++ geonode/layers/templates/upload/no_upload.html
@@ -5,14 +5,14 @@
 <div class="twocol">
     <p>{% trans "Your upload is either complete or you haven't resumed an earlier one." %}</p>
     <p>{% blocktrans %}Returning to the upload starting page in <span id="cnt">5</span>seconds {% endblocktrans %}.</p>
-    <p>{% blocktrans %} Or just go {% endblocktrans %}<a href="{% url "data_upload" %}">{% blocktrans %}now{% endblocktrans %}</a>.</p>
+    <p>{% blocktrans %} Or just go {% endblocktrans %}<a href="{% url "layer_upload" %}">{% blocktrans %}now{% endblocktrans %}</a>.</p>
 </div>
 <script type="text/javascript">
     var cnt = 5;
     window.setInterval(function() {
         Ext.get('cnt').dom.innerHTML = cnt--;
         if (cnt == 0) {
-            window.location = "{% url "data_upload" %}";
+            window.location = "{% url "layer_upload" %}";
         }
     }, 1000);
 </script>
