+++ geonode/templates/index.html
@@ -54,7 +54,7 @@
             {% blocktrans %}
               GeoNode enables you to upload, manage, and browse data layers. Search for data that is valuable to you, or upload your own data. 
             {% endblocktrans %}
-              <a href="{% url "data_upload" %}" class="btn btn-large">{% trans 'Upload Layers' %}</a>
+              <a href="{% url "layer_upload" %}" class="btn btn-large">{% trans 'Upload Layers' %}</a>
             {% blocktrans %}
               GeoNode enables you to compose and share maps. Create a map with our cartography tool, or explore maps shared by others. 
             {% endblocktrans %}
