+++ geonode/upload/views.py
@@ -59,11 +59,9 @@ _SESSION_KEY = 'geonode_upload_session'
 _ALLOW_TIME_STEP = getattr(settings, 'UPLOADER', False)
 if _ALLOW_TIME_STEP:
     _ALLOW_TIME_STEP = _ALLOW_TIME_STEP.get('OPTIONS', False).get('TIME_ENABLED', False)
-if any(settings.OGC_SERVER):
-    from geonode.utils import ogc_server_settings
-    _ASYNC_UPLOAD = True if ogc_server_settings and ogc_server_settings.DATASTORE else False
-else:
-    _ASYNC_UPLOAD = False
+
+from geonode.geoserver.helpers import ogc_server_settings
+_ASYNC_UPLOAD = True if ogc_server_settings and ogc_server_settings.DATASTORE else False
 
 # at the moment, the various time support transformations require the database
 if _ALLOW_TIME_STEP and not _ASYNC_UPLOAD:
